+++ sdk/storage/Azure.Storage.DataMovement.Blobs/tests/PageBlobDirectoryToBlockBlobDirectoryTests.cs
@@ -1,0 +1,111 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+extern alias DMBlobs;
+extern alias BaseBlobs;
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using Azure.Core;
+using BaseBlobs::Azure.Storage.Blobs;
+using BaseBlobs::Azure.Storage.Blobs.Models;
+using BaseBlobs::Azure.Storage.Blobs.Specialized;
+using Azure.Storage.DataMovement.Tests;
+using Azure.Storage.Test;
+using Azure.Storage.Test.Shared;
+using DMBlobs::Azure.Storage.DataMovement.Blobs;
+using NUnit.Framework;
+using Azure.Core.TestFramework;
+using Metadata = System.Collections.Generic.IDictionary<string, string>;
+using Azure.Storage.Shared;
+
+namespace Azure.Storage.DataMovement.Blobs.Tests
+{
+    [DataMovementBlobsClientTestFixture]
+    public class PageBlobDirectoryToBlockBlobDirectoryTests :
+        StartTransferBlobDirectoryCopyTestBase<PageBlobClient, BlockBlobClient>
+    {
+        public PageBlobDirectoryToBlockBlobDirectoryTests(
+            bool async,
+            object serviceVersion)
+        : base(async, serviceVersion)
+        {
+        }
+
+        protected override Task CreateObjectInDestinationAsync(
+            BlobContainerClient container,
+            long? objectLength = null,
+            string objectName = null,
+            Stream contents = null,
+            CancellationToken cancellationToken = default)
+            => CreateBlockBlobAsync(container, objectName, contents, cancellationToken);
+
+        protected override Task CreateObjectInSourceAsync(
+            BlobContainerClient container,
+            long? objectLength = null,
+            string objectName = null,
+            Stream contents = null,
+            TransferPropertiesTestType propertiesType = default,
+            CancellationToken cancellationToken = default)
+            => CreatePageBlobAsync(container, objectLength, objectName, contents, cancellationToken);
+
+        protected override async Task<IDisposingContainer<BlobContainerClient>> GetDestinationDisposingContainerAsync(
+            BlobServiceClient service = null,
+            string containerName = null,
+            CancellationToken cancellationToken = default)
+            => await DestinationClientBuilder.GetTestContainerAsync(service, containerName);
+
+        protected override StorageResourceContainer GetDestinationStorageResourceContainer(
+            BlobContainerClient containerClient,
+            string directoryPath,
+            TransferPropertiesTestType propertiesTestType = TransferPropertiesTestType.Default)
+        {
+            BlockBlobStorageResourceOptions options = default;
+            if (propertiesTestType == TransferPropertiesTestType.NewProperties)
+            {
+                options = new BlockBlobStorageResourceOptions(GetSetValuesResourceOptions());
+            }
+            else if (propertiesTestType == TransferPropertiesTestType.NoPreserve)
+            {
+                options = new BlockBlobStorageResourceOptions
+                {
+                    ContentDisposition = new(false),
+                    ContentLanguage = new(false),
+                    CacheControl = new(false),
+                    ContentType = new(false),
+                    Metadata = new(false)
+                };
+            }
+            else if (propertiesTestType == TransferPropertiesTestType.Preserve)
+            {
+                options = new BlockBlobStorageResourceOptions
+                {
+                    ContentDisposition = new(true),
+                    ContentLanguage = new(true),
+                    CacheControl = new(true),
+                    ContentType = new(true),
+                    Metadata = new(true)
+                };
+            }
+            return new BlobStorageResourceContainer(containerClient, new BlobStorageResourceContainerOptions()
+            {
+                BlobDirectoryPrefix = directoryPath,
+                BlobType = new(BlobType.Block),
+                BlobOptions = options
+            });
+        }
+
+        protected override StorageResourceContainer GetSourceStorageResourceContainer(BlobContainerClient containerClient, string directoryPath)
+            => new BlobStorageResourceContainer(containerClient, new BlobStorageResourceContainerOptions() { BlobDirectoryPrefix = directoryPath, BlobType = new(BlobType.Page) });
+
+        protected internal override BlockBlobClient GetDestinationBlob(BlobContainerClient containerClient, string blobName)
+            => containerClient.GetBlockBlobClient(blobName);
+
+        protected internal override PageBlobClient GetSourceBlob(BlobContainerClient containerClient, string blobName)
+            => containerClient.GetPageBlobClient(blobName);
+    }
+}