+++ sdk/storage/Azure.Storage.DataMovement.Files.Shares/BlobToFileSharesTests/ShareDirectoryToBlockBlobDirectoryTests.cs
@@ -5,65 +5,26 @@ extern alias DMBlob;
 extern alias BaseShares;
 
 using System;
-using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading;
 using System.Threading.Tasks;
 using Azure.Core;
-using Azure.Core.TestFramework;
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
 using Azure.Storage.Blobs.Specialized;
-using Azure.Storage.Common;
-using Azure.Storage.DataMovement.Files.Shares;
-using Azure.Storage.DataMovement.Tests;
 using BaseShares::Azure.Storage.Files.Shares;
-using BaseShares::Azure.Storage.Files.Shares.Models;
-using Azure.Storage.Test.Shared;
 using DMBlob::Azure.Storage.DataMovement.Blobs;
-using NUnit.Framework;
 
 namespace Azure.Storage.DataMovement.Blobs.Files.Shares.Tests
 {
     [BlobShareClientTestFixture]
-    public class ShareDirectoryToBlockBlobDirectoryTests : StartTransferDirectoryCopyTestBase
-        <ShareServiceClient,
-        ShareClient,
-        ShareClientOptions,
-        BlobServiceClient,
-        BlobContainerClient,
-        BlobClientOptions,
-        StorageTestEnvironment>
+    public class ShareDirectoryToBlockBlobDirectoryTests : StartTransferCopyFromShareDirectoryTestBase
     {
-        public const int MaxReliabilityRetries = 5;
-        private const string _fileResourcePrefix = "test-file-";
-        private const string _expectedOverwriteExceptionMessage = "BlobAlreadyExists";
-        protected readonly object _serviceVersion;
-
         public ShareDirectoryToBlockBlobDirectoryTests(
             bool async,
             object serviceVersion)
-            : base(async, _expectedOverwriteExceptionMessage, _fileResourcePrefix, null /* RecordedTestMode.Record /* to re-record */)
+            : base(async, serviceVersion)
         {
-            _serviceVersion = serviceVersion;
-            SourceClientBuilder = ClientBuilderExtensions.GetNewShareClientBuilder(Tenants, (ShareClientOptions.ServiceVersion)serviceVersion);
-            DestinationClientBuilder = ClientBuilderExtensions.GetNewBlobsClientBuilder(Tenants, (BlobClientOptions.ServiceVersion)serviceVersion);
-        }
-
-        protected override Task CreateDirectoryInDestinationAsync(BlobContainerClient destinationContainer, string directoryPath, CancellationToken cancellationToken = default)
-        {
-            CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
-            // No-op since blobs are virtual directories
-            return Task.CompletedTask;
-        }
-
-        protected override async Task CreateDirectoryInSourceAsync(ShareClient sourceContainer, string directoryPath, CancellationToken cancellationToken = default)
-        {
-            CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
-            ShareDirectoryClient directory = sourceContainer.GetRootDirectoryClient().GetSubdirectoryClient(directoryPath);
-            await directory.CreateIfNotExistsAsync(cancellationToken: cancellationToken);
         }
 
         protected override async Task CreateObjectInDestinationAsync(
@@ -97,7 +58,9 @@ namespace Azure.Storage.DataMovement.Blobs.Files.Shares.Tests
             ShareClient container,
             long? objectLength = null,
             string objectName = null,
-            Stream contents = null, CancellationToken cancellationToken = default)
+            Stream contents = null,
+            TransferPropertiesTestType propertiesType = default,
+            CancellationToken cancellationToken = default)
         {
             CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
             objectName ??= GetNewObjectName();
@@ -114,162 +77,44 @@ namespace Azure.Storage.DataMovement.Blobs.Files.Shares.Tests
             }
         }
 
-        protected override async Task<IDisposingContainer<BlobContainerClient>> GetDestinationDisposingContainerAsync(BlobServiceClient service = null, string containerName = null, CancellationToken cancellationToken = default)
-            => await DestinationClientBuilder.GetTestContainerAsync(service, containerName);
-
         protected override StorageResourceContainer GetDestinationStorageResourceContainer(
             BlobContainerClient sourceContainerClient,
             string directoryPath,
             TransferPropertiesTestType propertiesTestType = default)
-            => new BlobStorageResourceContainer(sourceContainerClient, new BlobStorageResourceContainerOptions() { BlobDirectoryPrefix = directoryPath, BlobType = new(BlobType.Block) });
-
-        protected override BlobContainerClient GetOAuthDestinationContainerClient(string containerName)
         {
-            BlobClientOptions options = DestinationClientBuilder.GetOptions();
-            BlobServiceClient oauthService = DestinationClientBuilder.GetServiceClientFromOauthConfig(Tenants.TestConfigOAuth, TestEnvironment.Credential, options);
-            return oauthService.GetBlobContainerClient(containerName);
-        }
-
-        protected override ShareClient GetOAuthSourceContainerClient(string containerName)
-        {
-            ShareClientOptions options = SourceClientBuilder.GetOptions();
-            options.ShareTokenIntent = ShareTokenIntent.Backup;
-            ShareServiceClient oauthService = SourceClientBuilder.GetServiceClientFromOauthConfig(Tenants.TestConfigOAuth, TestEnvironment.Credential, options);
-            return oauthService.GetShareClient(containerName);
-        }
-
-        protected override async Task<IDisposingContainer<ShareClient>> GetSourceDisposingContainerAsync(
-            ShareServiceClient service = null,
-            string containerName = null,
-            CancellationToken cancellationToken = default)
-            => await SourceClientBuilder.GetTestShareAsync(service, containerName);
-
-        protected override StorageResourceContainer GetSourceStorageResourceContainer(ShareClient containerClient, string directoryPath)
-        {
-            // Authorize with SAS when performing operations
-            if (containerClient.CanGenerateSasUri)
+            BlockBlobStorageResourceOptions options = default;
+            if (propertiesTestType == TransferPropertiesTestType.NewProperties)
             {
-                Uri sourceUri = containerClient.GenerateSasUri(BaseShares::Azure.Storage.Sas.ShareSasPermissions.All, Recording.UtcNow.AddDays(1));
-                ShareClient sasClient = InstrumentClient(new ShareClient(sourceUri, GetShareOptions()));
-                return new ShareDirectoryStorageResourceContainer(sasClient.GetDirectoryClient(directoryPath), default);
+                options = new BlockBlobStorageResourceOptions(GetSetValuesResourceOptions());
             }
-            return new ShareDirectoryStorageResourceContainer(containerClient.GetDirectoryClient(directoryPath), default);
-        }
-
-        protected override async Task VerifyEmptyDestinationContainerAsync(BlobContainerClient destinationContainer, string destinationPrefix, CancellationToken cancellationToken = default)
-        {
-            CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
-            IList<BlobItem> items = await destinationContainer.GetBlobsAsync(prefix: destinationPrefix, cancellationToken: cancellationToken).ToListAsync();
-            Assert.IsEmpty(items);
-        }
-
-        protected override async Task VerifyResultsAsync(
-            ShareClient sourceContainer,
-            string sourcePrefix,
-            BlobContainerClient destinationContainer,
-            string destinationPrefix,
-            TransferPropertiesTestType propertiesTestType = default,
-            CancellationToken cancellationToken = default)
-        {
-            CancellationHelper.ThrowIfCancellationRequested(cancellationToken);
-
-            // List all files in source blob folder path
-            List<string> sourceFileNames = new List<string>();
-
-            // Get source directory client and list the paths
-            ShareDirectoryClient sourceDirectory = string.IsNullOrEmpty(sourcePrefix) ?
-                sourceContainer.GetRootDirectoryClient() :
-                sourceContainer.GetDirectoryClient(sourcePrefix);
-            await foreach ((ShareDirectoryClient dir, ShareFileClient file) in ScanShareDirectoryAsync(
-                sourceDirectory, cancellationToken).ConfigureAwait(false))
+            else if (propertiesTestType == TransferPropertiesTestType.NoPreserve)
             {
-                if (file != default)
+                options = new BlockBlobStorageResourceOptions
                 {
-                    sourceFileNames.Add(file.Path.Substring(sourcePrefix.Length + 1));
-                }
+                    ContentDisposition = new(false),
+                    ContentLanguage = new(false),
+                    CacheControl = new(false),
+                    ContentType = new(false),
+                    Metadata = new(false)
+                };
             }
-
-            // List all files in the destination blob folder path
-            List<string> destinationFileNames = new List<string>();
-            await foreach (Page<BlobItem> page in destinationContainer.GetBlobsAsync(prefix: destinationPrefix, cancellationToken: cancellationToken).AsPages())
+            else if (propertiesTestType == TransferPropertiesTestType.Preserve)
             {
-                destinationFileNames.AddRange(page.Values.Select((BlobItem item) => item.Name.Substring(destinationPrefix.Length + 1)));
-            }
-
-            // Assert file and file contents
-            Assert.AreEqual(sourceFileNames.Count, destinationFileNames.Count);
-            sourceFileNames.Sort();
-            destinationFileNames.Sort();
-            for (int i = 0; i < sourceFileNames.Count; i++)
-            {
-                Assert.AreEqual(
-                    sourceFileNames[i],
-                    destinationFileNames[i]);
-
-                // Verify contents
-                string destinationFullName = string.Join("/", destinationPrefix, destinationFileNames[i]);
-                using Stream sourceStream = await sourceDirectory.GetFileClient(sourceFileNames[i]).OpenReadAsync(cancellationToken: cancellationToken);
-                using Stream destinationStream = await destinationContainer.GetBlobClient(destinationFullName).OpenReadAsync(cancellationToken: cancellationToken);
-                Assert.AreEqual(sourceStream, destinationStream);
-            }
-        }
-
-        public ShareClientOptions GetShareOptions()
-        {
-            var options = new ShareClientOptions((ShareClientOptions.ServiceVersion)_serviceVersion)
-            {
-                Diagnostics = { IsLoggingEnabled = true },
-                Retry =
+                options = new BlockBlobStorageResourceOptions
                 {
-                    Mode = RetryMode.Exponential,
-                    MaxRetries = MaxReliabilityRetries,
-                    Delay = TimeSpan.FromSeconds(Mode == RecordedTestMode.Playback ? 0.01 : 1),
-                    MaxDelay = TimeSpan.FromSeconds(Mode == RecordedTestMode.Playback ? 0.1 : 60)
-                },
-            };
-            if (Mode != RecordedTestMode.Live)
-            {
-                options.AddPolicy(new RecordedClientRequestIdPolicy(Recording), HttpPipelinePosition.PerCall);
+                    ContentDisposition = new(true),
+                    ContentLanguage = new(true),
+                    CacheControl = new(true),
+                    ContentType = new(true),
+                    Metadata = new(true)
+                };
             }
-
-            return InstrumentClientOptions(options);
-        }
-
-        private async IAsyncEnumerable<(ShareDirectoryClient Dir, ShareFileClient File)> ScanShareDirectoryAsync(
-            ShareDirectoryClient directory,
-            [EnumeratorCancellation] CancellationToken cancellationToken)
-        {
-            Argument.AssertNotNull(directory, nameof(directory));
-
-            Queue<ShareDirectoryClient> toScan = new();
-            toScan.Enqueue(directory);
-
-            while (toScan.Count > 0)
+            return new BlobStorageResourceContainer(sourceContainerClient, new BlobStorageResourceContainerOptions()
             {
-                ShareDirectoryClient current = toScan.Dequeue();
-                await foreach (ShareFileItem item in current.GetFilesAndDirectoriesAsync(
-                    cancellationToken: cancellationToken).ConfigureAwait(false))
-                {
-                    if (item.IsDirectory)
-                    {
-                        ShareDirectoryClient subdir = current.GetSubdirectoryClient(item.Name);
-                        toScan.Enqueue(subdir);
-                        yield return (Dir: subdir, File: null);
-                    }
-                    else
-                    {
-                        yield return (Dir: null, File: current.GetFileClient(item.Name));
-                    }
-                }
-            }
-        }
-
-        [Test]
-        [LiveOnly] // https://github.com/Azure/azure-sdk-for-net/issues/33082
-        public override Task DirectoryToDirectory_OAuth()
-        {
-            // NoOp this test since File To Blob
-            return Task.CompletedTask;
+                BlobDirectoryPrefix = directoryPath,
+                BlobType = new(BlobType.Block),
+                BlobOptions = options
+            });
         }
     }
 }